Update the test logic
1. v1 (previous commit) needs to clone linux-next on Windows client and
use this as "controller" role to find out the middle commit by git
bisect, then pass this commit ID to VM to ask VM to reset tot this
commit, then build the kernel and test the performance. But, sometimes,
the Commit ID found by Windows linux-next git bisect will not be found
from the VM's linux-next source tree. So, update the logic to below
(this commit).
2. v2 (this commit) will do:
a) clone linux-next on one VM, then drive git bisect to find the middle
commit and build the kernel.
b) copy the new kernel (*.deb if Ubuntu) to another VM
c) install the new kernel on both VMs
d) reboot to detect the new kernel is boouup or not
* if new kernel boot failed, then run git bisect skip this commit
* if bott up successfully, then run test
e) according to the test result, tell git bisect the result (good or
bad) then run the next git bisect test.