Fix hard coded log folder path
Changed the way the $str string is formed (in SummaryToString and SendEmail
methods) by adding a new input variable, $logDir, for those functions. The
variable is in fact the path that RunTests method, from lisa.ps1, uses in
order to create the log folder.
The script tests if the path is absolute, if not it gets the current
absolute path from the system, and forms the final string that will be
printed in the ica.log file or inserted in the email to be sent.