Configure tox.ini for use in CI
In order to automate a CI process for future development of
the parser a setup.py was created for the project and
a tox is used to run tests under multiple environments.

Tests are run under a python27 env and also pep8 in order
to check the code format.

Project main files were moved in a separate folder for
better logic separation.

Changes were made to current unit tests in order to work with tox.