Add test cases to check MAX NIC support
Four new test cases are added to check if the VM can handle
the maximum amount of NICs that Hyper-V can attach to it - 8 synthetic
and 4 legacy, at the moment.

The first 3 test cases, MaxSyntheticNIC, MaxLegacyNIC, MaxNIC,
use the NET_ADD_MAX_NIC.ps1 setup script to add 7, 4 or 11 new
NICs, depending on each case. They follow the same logic:
	1. Check if VM sees the same amount of interfaces as the ones added
	during setup
	2. Configure each interface, using dhcp, bring each one up and
	check if they have an IP address
	3. Check external connection for each interface, after each
	configuration and also after all interfaces are up
	4. Compare the IP address values from KVP with the ones on the VM

HotAddRemoveMaxNIC is targeted only to Gen 2 VMs. It uses the same logic
of HotAddRemoveNIC, only with the maximum amount of NICs, that is
permitted by Hyper-V.