Added extra config to cover other distros
A new set of config steps were added so
that the automation can run on rhel/centos
and ubuntu distros.

For rhel/centos the same method of using
the rc.local file, to run a script on startup,
was used. For ubuntu the script is triggered
by a cron job on reboot (rc.local config
was not properly working on some versions)

Also redundant code was removed from the
bash script.