Dynamic Memory - New TCs, code cleanup & improvements
-Separate Hot Add tests into 3 categories:
---Basic Hot Add tests – Tests which use only one VM – Adding 128MB
chunks every 10 seconds. If these tests fail, then we know it’s an issue
with the test VM.
---Complex Hot Add tests – Tests which use 2 VMs – Adding 128MB chunks
every 10 seconds. These tests may fail due to other VMs
running/starting/stopping on the host or problems with the Dependency
VM.
---Stress Hot Add tests – 2 new TCs. One will add 512MB chunks every 5
seconds and the other will add 512MB chunks every second. This will
probably cause failures on WS2012 and WS2012R2 but may pass on WS2016
-Fixed an issue where an error occured on DM_Configure_Memory.ps1 when
allocating memory to the VM
-Moved duplicate code to TCUtils.ps1
-Code cleanup