Save VM output to log file
A new feature was added in order to save the output
of the tested VMs in separate log files.
Output log files are created for each VM used and
saved in the test suite log path, under the following
name format "${vmName}-icaserial.log"

Also the CreateVMs setup script was update in order
to configure the com2 port with an unique path.