Changed TargetIP and IQN handling in pass_offline
This commit changes the way STOR_pass_offline handles the TargetIP
and IQN parameters.
More specifically if those are not present inside the xml file the
test will be skipped and the logs will be set accordingly.
The options are now commented by default in the STOR_VHD xml file
so the test will be skipped.