Fixed the init image check
This commit fixes the way modules_check.sh checks for the init
image. Instead of having one 'IF' structure with multiple conditions
chained with logical 'OR' operators, the check is done using
multitple 'IF' instructions.
This prevents the error showing up in the log, in case the first
condition isn't true.