New-PSSession: Work around "Kerberos Double Hop" issue
We currently use kerberos authentication in AD, If LISA is run inside a
PSSession, creating a new pssession in the test would require delegation
configured (and even it is configured, double hop to localhost is not
allowed). This patch use fresh identity to create PSSession instead of
using delegation (only if $env:HostUser and $env:HostPassword are defined,
otherwise the behaviour is not changed).