Quick fix for reverting snapshot issue
When running the full test suite additional snapshots get created
enven if they already exist. This causes the Restote-VMSnapshot
commandlet to receive a System.Object[], list that contains more
than a a single snapshot. This, in turn, causes all the tests that
use the RevertSnapshot.ps1 setup script to fail.

This commit fixes that issue.